posts/ProLogiQA/pyproject.toml
[project]
name = "prologiqa"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "ty>=0.0.13",
]
